+++ package.json
@@ -41,6 +41,7 @@
     "chai": "^4.3.6",
     "chai-http": "^4.3.0",
     "cookie-parser": "^1.4.6",
+    "cors": "^2.8.5",
     "coverage": "^0.4.1",
     "cross-env": "^7.0.3",
     "dotenv": "^16.0.1",
@@ -48,6 +49,8 @@
     "husky": "^8.0.1",
     "jsonwebtoken": "^8.5.1",
     "mocha": "^10.0.0",
+    "morgan": "^1.10.0",
+    "node": "^17.7.2",
     "nodemon": "^2.0.16",
     "nyc": "^15.1.0",
     "passport": "^0.5.3",
